Change CTF event format
Now CTF event description is created dynamically based in the amount of
CPUs.
Also the format of the field was changed in order to use floating point
representation instead int64.

+++ plugins/tracers/gstcpuusagecompute.h
@@ -34,7 +34,7 @@ G_BEGIN_DECLS
 {
   /* CPU core number */
   gint cpu_num;
-  gdouble cpu_load[CPU_NUM_MAX];
+  gfloat cpu_load[CPU_NUM_MAX];
 
   gint user[CPU_NUM_MAX];       /* Time spent in user mode */
   gint user_aux[CPU_NUM_MAX];   /* Time spent in user mode */
